annotated/requirements.txt
flask
psutil
python-dotenv
pytest
flask-migrate
gunicorn
flask-sqlalchemy
flask-login
flask-mail
twilio
celery
redis
psycopg2-binary
bcrypt
pandas
apscheduler
flask-cors
